+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "webwithpy"
-version = "0.2"
+version = "0.3"
 description = ""
 authors = ["Sven Keimpema"]
 readme = "README.md"
@@ -8,10 +8,7 @@ packages = [{include = "webwithpy/**/*.*", format = "sdist"}]
 
 [tool.poetry.dependencies]
 python = "^3.7"
-starlette = "^0.27.0"
-jinja2= "^3.1.2"
-aiohttp = "^3.8.1"
-aiofiles = "^23.2.1"
+asyncio = "^3.4.3"
 
 [build-system]
 requires = ["poetry-core"]
