+++ webwithpy/orm/db.py
@@ -16,15 +16,22 @@ def dict_factory(cursor, row):
 class DB:
     conn = None
     cursor = None
+    driver = None
     tables = None
 
     def __init__(self, db_path: Union[Path, str]):
-        if DB.conn is None and DB.cursor is None:
-            if isinstance(db_path, str):
-                db_path = Path(db_path)
+        driver, db_path = str(db_path).split('/')
 
-            # make sure the db exists
-            db_path.touch()
+        if DB.driver is None:
+            DB.driver = self._get_driver(driver)
+
+        if DB.conn is None and DB.cursor is None:
+            db_path = Path(db_path)
+
+            # make sure the folders to the db and the db itself exists
+            db_folder_path = Path(str(db_path)[:-len(str(db_path.name))])
+            db_folder_path.mkdir(parents=True, exist_ok=True)
+            db_path.touch(exist_ok=True)
 
             DB.conn = sqlite.connect(db_path)
             DB.conn.row_factory = dict_factory
@@ -65,11 +72,21 @@ class DB:
             DB.tables[table_name] = tbl
             self._create_table(table_name, *[field for field in table_fields.values()])
 
+    def _get_driver(self, driver: str):
+        match driver:
+            case 'sqlite':
+                from .drivers.sqlite import SqliteDriver
+                from .dialects.sqlite import SqliteDialect
+                return SqliteDriver(self.table)
+        return ""
+
     def _create_table(self, table_name: str, *fields: Field):
         sql = f"CREATE TABLE IF NOT EXISTS {table_name} ("
+
         for field in fields:
             sql += f"{field.field_name} {field.field_type}, "
         sql = sql[:-2] + ")"
+
         DB.cursor.execute(sql)
 
     def __getattribute__(self, item):
